pyproject.toml
[tool.poetry]
name = "fastapi-mapproxy"
version = "0.1.0"
description = ""
authors = ["Francesco Bartoli <francesco.bartoli@geobeyond.it>"]
readme = "README.md"
packages = [
    {include = "mapproxy"},
    {include = "mapproxy/**/*.py"},
]

[tool.poetry.dependencies]
python = "^3.10"
mapproxy = {git = "https://github.com/mapproxy/mapproxy.git", branch = "master"}
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic-settings = "^2.6.1"
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
